+++ workspace/solutions/ecmascript/src/s0/s3/s0/s3/s_0303_test.ts
@@ -1,0 +1,11 @@
+import test from "bun:test";
+
+import { E } from "@repo/common";
+
+import { t } from "@/utils/test";
+
+t(
+    test,
+    import.meta.dir,
+    E.SolutionFormat.class,
+);