+++ workspace/solutions/go/s0/s0/s0/s1/s_0001_test.go
@@ -9,7 +9,7 @@ import (
 )
 
 func TestSolutions(t *testing.T) {
-	utils.Test(
+	utils.T(
 		[]any{
 			s_0001_00.Solution,
 		},