+++ workspace/solutions/java/lib/src/test/java/s0/s0/s0/s1/s_0001_test.java
@@ -2,15 +2,15 @@ package s0.s0.s0.s1;
 
 import org.junit.jupiter.api.Test;
 import s0.s0.s0.s1.s_0001_00.s_0001_00;
-import utils.Test_;
+import utils.T;
 
 class s_0001_test {
     @Test
     void test() {
-        Test_ test = new Test_(
+        T t = new T(
             this.getClass(),
             s_0001_00.class
         );
-        test.run();
+        t.run();
     }
 }